adds string_tensor.ts file, adds reshape
