+++ src/exports.ts
@@ -34,7 +34,7 @@ import {categoricalCrossentropy, cosineProximity, meanAbsoluteError, meanAbsolut
 import {binaryAccuracy, binaryCrossentropy, categoricalAccuracy} from './metrics';
 import {loadModelInternal, Sequential, SequentialConfig} from './models';
 import {OneHot, OneHotLayerConfig} from './preprocess-layers/preprocess_core';
-export * from './preprocess-layers/preprocess_core';
+export * from './preprocess-layers/string_tensor';
 import {l1, L1Config, L1L2, L1L2Config, l2, L2Config, Regularizer} from './regularizers';
 import {SymbolicTensor} from './types';
